feature: change user_file structure & add share api

+++ maid-cassandra/src/main/java/org/dreamcat/maid/cassandra/service/UserServiceImpl.java
@@ -1,35 +1,47 @@
 package org.dreamcat.maid.cassandra.service;
 
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.dreamcat.common.web.asm.BeanCopierUtil;
 import org.dreamcat.common.web.core.RestBody;
 import org.dreamcat.common.web.exception.NotFoundException;
-import org.dreamcat.common.web.util.BeanCopierUtil;
 import org.dreamcat.maid.api.controller.user.UpdateUserQuery;
 import org.dreamcat.maid.api.controller.user.UserView;
-import org.dreamcat.maid.api.service.FileOpService;
 import org.dreamcat.maid.api.service.UserService;
 import org.dreamcat.maid.cassandra.dao.AvatarDao;
 import org.dreamcat.maid.cassandra.dao.UserDao;
+import org.dreamcat.maid.cassandra.dao.UserFileDao;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ServerWebExchange;
+
+import static org.dreamcat.maid.cassandra.core.RestCodes.avatar_not_found;
+import static org.dreamcat.maid.cassandra.core.RestCodes.user_not_found;
 
 /**
  * Create by tuke on 2020/3/18
  */
+@Slf4j
 @RequiredArgsConstructor
 @Service
 public class UserServiceImpl implements UserService {
     private final UserDao userDao;
     private final AvatarDao avatarDao;
+    private final UserFileDao userFileDao;
     private final CommonService commonService;
-    private final FileOpService fileOpService;
+    private final FileBatchService fileBatchService;
 
     @Override
     public RestBody<?> deleteUser(ServerWebExchange exchange) {
-        fileOpService.removeNumerous("/", exchange);
         var uid = commonService.retrieveUid(exchange);
-        userDao.deleteById(uid);
+        var user = userDao.findById(uid).orElse(null);
+        if (user == null) {
+            return RestBody.error(user_not_found, "user not found");
+        }
+
+        fileBatchService.deleteAllShareFile(uid);
+        userFileDao.deleteByUid(uid);
         avatarDao.delete(uid);
+        userDao.deleteById(uid);
         return RestBody.ok();
     }
 
@@ -38,7 +50,7 @@ public class UserServiceImpl implements UserService {
         var uid = commonService.retrieveUid(exchange);
         var user = userDao.findById(uid).orElse(null);
         if (user == null) {
-            throw new NotFoundException("User doesn't exist, maybe deleted already");
+            return RestBody.error(user_not_found, "user not found");
         }
 
         var view = BeanCopierUtil.copy(user, UserView.class);
@@ -51,7 +63,7 @@ public class UserServiceImpl implements UserService {
         var uid = commonService.retrieveUid(exchange);
         var avatar = avatarDao.find(uid);
         if (avatar == null) {
-            throw new NotFoundException("Avatar doesn't exist");
+            throw new NotFoundException(avatar_not_found, "avatar not found");
         }
         return avatar.getAvatar();
     }
@@ -61,7 +73,7 @@ public class UserServiceImpl implements UserService {
         var uid = commonService.retrieveUid(exchange);
         var user = userDao.findById(uid).orElse(null);
         if (user == null) {
-            throw new NotFoundException("User doesn't exist, maybe deleted already");
+            return RestBody.error(user_not_found, "user not found");
         }
 
         BeanCopierUtil.copy(query, user);
